refactor:Dto 구분,update:친구 삭제 요청 API 인자
