+++ src/main/java/link/sendwish/backend/controller/FriendController.java
@@ -1,10 +1,12 @@
 package link.sendwish.backend.controller;
 
 
-import link.sendwish.backend.auth.TokenInfo;
 import link.sendwish.backend.common.exception.DtoNullException;
 import link.sendwish.backend.dtos.*;
-import link.sendwish.backend.entity.Member;
+import link.sendwish.backend.dtos.friend.FriendAddRequestDto;
+import link.sendwish.backend.dtos.friend.FriendAddResponseDto;
+import link.sendwish.backend.dtos.friend.FriendDeleteRequestDto;
+import link.sendwish.backend.dtos.friend.FriendResponseDto;
 import link.sendwish.backend.service.MemberService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -25,7 +27,7 @@ public class FriendController {
             if(nickname == null){
                 throw new DtoNullException();
             }
-            List<MemberFriendResponseDto> dtos = memberService.findFriendsByMember(nickname);
+            List<FriendResponseDto> dtos = memberService.findFriendsByMember(nickname);
             return ResponseEntity.ok().body(dtos);
         } catch (Exception e) {
             e.printStackTrace();
@@ -37,12 +39,12 @@ public class FriendController {
     }
 
     @PostMapping("/friend")
-    public ResponseEntity<?> addFriend(@RequestBody MemberFriendAddRequestDto dto) {
+    public ResponseEntity<?> addFriend(@RequestBody FriendAddRequestDto dto) {
         try {
             if (dto.getMemberNickname() == null || dto.getAddMemberNickname() == null) {
                 throw new DtoNullException();
             }
-            MemberFriendAddResponseDto dtos = memberService.addFriendToMe(dto);
+            FriendAddResponseDto dtos = memberService.addFriendToMe(dto);
             return ResponseEntity.ok().body(dtos);
         } catch (Exception e) {
             e.printStackTrace();
@@ -53,14 +55,13 @@ public class FriendController {
         }
     }
 
-    @DeleteMapping("/friend/{nickname}/{friendNickname}")
-    public ResponseEntity<?> deleteFriend(@PathVariable("nickname") String nickname,
-                                          @PathVariable("friendNickname") String friendNickname){
+    @DeleteMapping("/friend")
+    public ResponseEntity<?> deleteFriend(@RequestBody FriendDeleteRequestDto dto) {
         try{
-            if(nickname == null || friendNickname == null){
+            if(dto.getNickname() == null || dto.getFriendNickname() == null){
                 throw new DtoNullException();
             }
-            memberService.deleteFriend(nickname, friendNickname);
+            memberService.deleteFriend(dto.getNickname(), dto.getFriendNickname());
             return ResponseEntity.ok().body("친구 삭제 성공");
         } catch (Exception e) {
             e.printStackTrace();
