+++ src/main/java/link/sendwish/backend/controller/ItemController.java
@@ -2,6 +2,9 @@ package link.sendwish.backend.controller;
 
 import link.sendwish.backend.common.exception.DtoNullException;
 import link.sendwish.backend.dtos.*;
+import link.sendwish.backend.dtos.item.ItemCreateRequestDto;
+import link.sendwish.backend.dtos.item.ItemEnrollmentRequestDto;
+import link.sendwish.backend.dtos.item.ItemResponseDto;
 import link.sendwish.backend.entity.Collection;
 import link.sendwish.backend.entity.Item;
 import link.sendwish.backend.entity.Member;
