+++ src/main/java/link/sendwish/backend/dtos/collection/CollectionDeleteRequestDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.collection;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,8 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class CollectionUpdateRequestDto {
+public class CollectionDeleteRequestDto {
     private String nickname;
-    private String newTitle;
     private Long collectionId;
 }
