+++ src/main/java/link/sendwish/backend/dtos/friend/FriendAddRequestDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.friend;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberFriendAddRequestDto {
+public class FriendAddRequestDto {
     private String memberNickname; // 친구를 추가할 본인의 memberNickname
     private String addMemberNickname; // 추가할 친구의 memberNickname
 }
