+++ src/main/java/link/sendwish/backend/dtos/friend/FriendAddResponseDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.friend;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberFriendAddResponseDto {
+public class FriendAddResponseDto {
     private String myNickname; // my nickname
     private String friendNickname; // 친구의 nickname
 }
