+++ src/main/java/link/sendwish/backend/dtos/friend/FriendDeleteRequestDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.friend;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class MemberFriendResponseDto {
-    private Long friend_id;
-    private String friend_nickname;
+public class FriendDeleteRequestDto {
+    private String nickname; // 내 닉네임
+    private String friendNickname;
 }
