+++ src/main/java/link/sendwish/backend/dtos/member/MemberRequestDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
