+++ src/main/java/link/sendwish/backend/dtos/member/MemberResponseDto.java
@@ -1,4 +1,4 @@
-package link.sendwish.backend.dtos;
+package link.sendwish.backend.dtos.member;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
-public class ItemCreateRequestDto {
-    private String url;
+public class MemberResponseDto {
+    private Long id;//mypage
     private String nickname;
 }
